Introduced a unified logging daemon, logd.
logd is a new daemon, which reads from a file (which may be a pipe) and writes
all read data to a log file. This removes the duplication between syslogd and
klogd: now, syslogd reads from a Unix socket and writes to a FIFO, while klogd
is no longer needed, thanks to logd and /proc/kmsg.