Made relayd adhere to the Unix philosophy.
Now, it is split into relayd (a generic server) and client (a generic TCP
client). The former listens on a given port and runs the latter, with a socket
as its standard input and output pipes.

In addition, the blacklist consists of regular expressions and traffic
analyzation is a lot faster now.