XORed the system log.
The system log is easy to modify because it's textual. Protecting it using strong encryption or compression doesn't make much sense, because it's always possible to use a scripting language (e.g Python) one-liner or a special tool to modify the log.

XOR is good enough to prevent trivial tampering attempts.